Rotus: Player 피격 시스템 업로드
-체력을 델타타임을 이용하여 감소시키기 위해 int에서 float형으로 변환

+++ SurvivorsLike/Assets/SurvivorsLike/Source/GameManager.cs
@@ -21,8 +21,8 @@ public class GameManager : MonoBehaviour
     public int Exp;
     public int Level;
     public int KillCount;
-    public int MaxHP = 100;
-    public int HP;
+    public float MaxHP = 100;
+    public float HP;
 
     void Awake()
     {
